organized database inspectors into seperate directory
