+++ tests/migration_engine/test_db_inspectors.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import shutil
 import pytest
 from skibidi_orm.migration_engine.config import SQLite3Config
-from skibidi_orm.migration_engine.db_inspector import SqliteInspector
+from skibidi_orm.migration_engine.db_inspectors.db_inspector import SqliteInspector
 import sqlite3
